refactor(codegen): extrai analysis/module_globals.rs de func.rs
Move collect_module_globals + collect_idents_referenced_in_user_fns
(351 linhas) para crates/rts-codegen/src/codegen/lower/analysis/
module_globals.rs. Detecta vars top-level que viram data globais
(referenciadas por user fns) vs vars promovidas a Variables locais
em __RTS_MAIN.

func.rs: 8524 -> 8175 linhas. Suite verde.
